solution/requirements.txt
torch>=2.1
torchvision>=0.16
numpy
Pillow
tqdm